+++ examples/rxJavaExample/src/main/java/io/realm/examples/rxjava/retrofit/GitHubUser.java
@@ -19,9 +19,9 @@ package io.realm.examples.rxjava.retrofit;
 /**
  * Model class for GitHub users: https://developer.github.com/v3/users/#get-a-single-user
  */
-public class GitHubUser {
+@SuppressWarnings("unused")
+class GitHubUser {
     public String name;
-    public String email;
     public int public_repos;
     public int public_gists;
 }