+++ realm/realm-library/src/main/java/io/realm/DynamicRealm.java
@@ -22,8 +22,8 @@ import io.realm.exceptions.RealmException;
 import io.realm.exceptions.RealmIOException;
 import io.realm.internal.Table;
 import io.realm.internal.TableView;
-import rx.Observable;
 import io.realm.internal.log.RealmLog;
+import rx.Observable;
 
 /**
  * DynamicRealm is a dynamic variant of {@link io.realm.Realm}. This means that all access to data and/or queries are
@@ -165,7 +165,7 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Get all objects of a specific class name.
+     * Gets all objects of a specific class name.
      *
      * @param className the Class to get objects of.
      * @return a {@link RealmResults} list containing the objects. If no results where found, an empty list
@@ -177,8 +177,9 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Get all objects of a specific class name sorted by a field. If no objects exist, the returned
-     * {@link RealmResults} will not be {@code null}. Use {@link RealmResults#size()} to check the number of objects instead.
+     * Gets all objects of a specific class name sorted by a field. If no objects exist, the returned
+     * {@link RealmResults} will not be {@code null}. Use {@link RealmResults#size()} to check the number of objects
+     * instead.
      *
      * @param className the class to get all objects from.
      * @param fieldName the field name to sort by.
@@ -200,9 +201,9 @@ public final class DynamicRealm extends BaseRealm {
 
 
     /**
-     * Get all objects of a specific class name sorted by two specific field names.  If no objects exist,
-     * the returned {@link RealmResults} will not be {@code null}. Use {@link RealmResults#size()} to check the number of
-     * objects instead.
+     * Gets all objects of a specific class name sorted by two specific field names.  If no objects exist,
+     * the returned {@link RealmResults} will not be {@code null}. Use {@link RealmResults#size()} to check the number
+     * of objects instead.
      *
      * @param className the class to get all objects from.
      * @param fieldName1 the first field name to sort by.
@@ -221,7 +222,7 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Get all objects of a specific class name sorted by multiple fields.  If no objects exist, the
+     * Gets all objects of a specific class name sorted by multiple fields.  If no objects exist, the
      * returned {@link RealmResults} will not be {@code null}. Use {@link RealmResults#size()} to check the number of
      * objects instead.
      *
@@ -252,7 +253,7 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Return a distinct set of objects of a specific class. As a Realm is unordered, it is undefined which objects are
+     * Returns a distinct set of objects of a specific class. As a Realm is unordered, it is undefined which objects are
      * returned in case of multiple occurrences.
      *
      * @param className the Class to get objects of.
@@ -270,7 +271,7 @@ public final class DynamicRealm extends BaseRealm {
     }
 
     /**
-     * Return a distinct set of objects of a specific class. As a Realm is unordered, it is undefined which objects are
+     * Returns a distinct set of objects of a specific class. As a Realm is unordered, it is undefined which objects are
      * returned in case of multiple occurrences.
      * This method is only available from a Looper thread.
      *