+++ realm/realm-library/src/main/java/io/realm/internal/TableOrView.java
@@ -343,7 +343,7 @@ public interface TableOrView {
     TableQuery where();
 
     /**
-     * Find a row with in the table or view with the given index.
+     * Finds a row with in the table or view with the given index.
      *
      * @param rowIndex the index of the row.
      * @return the index if found, or -1 for not found.